web3/vyper/cu-int-py-and-vy/sec-2-web3py/pyproject.toml
[project]
name = "sec-2-web3py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "vyper>=0.4.0",
    "vyper-lsp>=0.1.0",
    "web3>=6.20.3",
]
